navigate to tabs view
