+++ app/_layout.tsx
@@ -28,7 +28,7 @@ export default function RootLayout() {
 	return (
 		<Stack>
 			<Stack.Screen name='index' options={{ headerShown: false }} />
-			<Stack.Screen name='tabs' options={{ headerShown: false }} />
+			<Stack.Screen name='root' options={{ headerShown: false }} />
 			<Stack.Screen name='+not-found' />
 		</Stack>
 	);
