+++ app/root/_layout.tsx
@@ -1,0 +1,9 @@
+import { Stack } from 'expo-router';
+
+export default function RootLayout() {
+	return (
+		<Stack>
+			<Stack.Screen name='tabs' options={{ headerShown: false }} />
+		</Stack>
+	);
+}
